+++ src/main/java/com/tagnumelite/projecteintegration/addons/PsiAddon.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019-2021 TagnumElite
+ * Copyright (c) 2019-2023 TagnumElite
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -24,7 +24,7 @@ package com.tagnumelite.projecteintegration.addons;
 
 import com.tagnumelite.projecteintegration.api.recipe.ARecipeTypeMapper;
 import moze_intel.projecte.api.mapper.recipe.RecipeTypeMapper;
-import net.minecraft.item.crafting.RecipeType;
+import net.minecraft.world.item.crafting.RecipeType;
 import vazkii.psi.common.crafting.ModCraftingRecipes;
 import vazkii.psi.common.crafting.recipe.TrickRecipe;
 
@@ -40,7 +40,7 @@ public class PsiAddon {
 
         @Override
         public boolean canHandle(RecipeType<?> recipeType) {
-            returnrecipeType == ModCraftingRecipes.TRICK_RECIPE_TYPE;
+            return recipeType == ModCraftingRecipes.TRICK_RECIPE_TYPE;
         }
     }
 }